backend/requirements.txt
# Requirements for Mind Wellness Backend
alembic==1.14.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
stripe==11.2.0
tomli==2.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
